pyproject.toml
[project]
name = "asyncroscopy"
version = "0.1.0"
description = "Asyncroscopy: asynchronous control framework for STEM automation"
readme = "README.md"
requires-python = ">3.11"
dependencies = [
    # Other project dependencies
    "colorama>=0.4.6",
    "defusedxml>=0.7.1",
    "gevent>=25.9.1",
    "h5py>=3.15.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "multiprocess>=0.70.19",
    "numpy>=2.3.5",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "pytango==10.1.2",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
    "sidpy>=0.12.8",
]

[project.optional-dependencies]
autoscript = [
    "autoscript-core",
    "autoscript-tem-microscope-client",
    "autoscript-tem-microscope-client-tests",
    "autoscript-tem-snippets-tests",
    "autoscript-tem-toolkit",
    "thermoscientific-logging",
]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
include = ["asyncroscopy*"]

[tool.uv.sources]
autoscript-core = { path = "stubs/autoscript_core-1.0.0-py3-none-any.whl" }
autoscript-tem-microscope-client-tests = { path = "stubs/autoscript_tem_microscope_client_tests-1.15.0-py3-none-any.whl" }
autoscript-tem-microscope-client = { path = "stubs/autoscript_tem_microscope_client-1.15.0-py3-none-any.whl" }
autoscript-tem-snippets-tests = { path = "stubs/autoscript_tem_snippets_tests-1.15.0-py3-none-any.whl" }
autoscript-tem-toolkit = { path = "stubs/autoscript_tem_toolkit-1.15.0-py3-none-any.whl" }
thermoscientific-logging = { path = "stubs/thermoscientific_logging-1.0.0-py3-none-any.whl" }
